% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_mat_pval.R
\name{adjust_mat_pval}
\alias{adjust_mat_pval}
\title{adjust_mat_pval}
\usage{
adjust_mat_pval(mat, method = "BH", out_type = "matrix")
}
\arguments{
\item{mat}{numeric matrix}

\item{method}{multiple testing correction method. default 'BH'}

\item{out_type}{'matrix' or 'dataframe'. matrix returns symmetrical matrix.
dataframe returns long dataframe}
}
\value{
Adjusted p-values as either matrix (default) or long data.frame
}
\description{
Adjust matrix of p-values for multiple correction. Helpful for correlation matrices.
}
\examples{
# load example data
data(dss_example)

# subset features, features in columns
feat_mat <- dss_example$merged_abundance_id[1:6,2:26]
rownames(feat_mat) <- dss_example$merged_abundance_id[1:6,1]
feat_mat <- t(feat_mat)

# correlation matrix
corr_result <- psych::corr.test(feat_mat)

adjust_mat_pval(corr_result$p.value)
adjust_mat_pval(corr_result$p.value, out_type='dataframe')
}
