% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_diversity.R
\name{alpha_diversity}
\alias{alpha_diversity}
\title{alpha_diversity}
\usage{
alpha_diversity(count_df)
}
\arguments{
\item{count_df}{Count table with samples in columns, taxa in rows}
}
\value{
data frame of alpha diversity metrics for each sample
}
\description{
Calculates alpha diversity metrics from count table: Shannon's D,
Simpson's index, evenness, richness. Calculations are based on
Jost et al 2006 https://doi.org/10.1111/j.2006.0030-1299.14714.x
}
\examples{
# get example data
data(asv_example)

# rownames have to be features
asv_counts <- data.frame(asv_example[2:ncol(asv_example)], row.names=asv_example$sequence)
alpha_div <- alpha_diversity(asv_counts)

}
