% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_example.R
\docType{data}
\name{asv_example}
\alias{asv_example}
\title{Example ASV counts table}
\format{
dataframe
}
\usage{
data(asv_example)
}
\description{
As subset of data from a V3-V4 16S rRNA human dataset
}
\keyword{datasets}
