% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcdissimilarity.R
\name{bcdissimilarity}
\alias{bcdissimilarity}
\title{bcdissimilarity}
\usage{
bcdissimilarity(relab_mat, met, sample_id, var)
}
\arguments{
\item{relab_mat}{data.frame or matrix. relative abundance table with samples in
columns and features in rows; feature IDs as rownames}

\item{met}{dataframe. metadata table}

\item{sample_id}{string. Sample identifier (column name in `met`)}

\item{var}{string. Metadata variable (column name in `met`) that specifies groups.
set to NULL to bypass between/within group assignment}
}
\value{
list.
         `bc_df` is long dataframe of samples being compared and corresponding distances.
         `bc_dist` matrix of bray curtis distances
}
\description{
Calculate Bray-Curtis Dissimilarity for samples within and between groups.
}
\examples{

data(dss_example)

count_table <- dss_example$merged_abundance_id \%>\% column_to_rownames('featureID')
relab_table <- relab(count_table)

met_table <- dss_example$metadata

# bray curtis for one metadata variable
bc_result <- bcdissimilarity(relab_table, met_table, 'sampleID','Phenotype')

pdata <- bc_result$bc_df
p_phen <- ggplot(pdata, aes(x=value, y=dist)) +
   geom_violin() +
   theme_bw(14) +
   ylab('Bray-Curtis Dissimilarity') +
   xlab('Phenotype')

for multiple metadata variables
bc_data <- c()
for(var in c('Phenotype','Genotype')) {
    bc_result <- bcdissimilarity(relab_table, met_table, 'sampleID', var)
    bc_data <- rbind(bc_data, bc_result$bc_df)
}

p_bc <- ggplot(bc_data, aes(x=comparison, y=dist)) +
   geom_violin() +
   theme_bw(14) +
   facet_wrap(~met_var) +
   ylab('Bray-Curtis Dissimilarity') +
   theme(axis.title.x=element_blank())

}
