% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cor_ci.R
\name{compare_cor_ci}
\alias{compare_cor_ci}
\title{compare_cor_ci}
\usage{
compare_cor_ci(
  mat,
  group,
  method = "pearson",
  adjust = "BH",
  alternative = "two.sided",
  conf_level = 0.95
)
}
\arguments{
\item{mat}{matrix with features in columns.
correlations calculated between feature pairs for each group}

\item{group}{group assignment of samples corresponding to rows in mat.}

\item{method}{correlation method pearson, spearman, kendall}

\item{adjust}{multiple testing correction method.
BH, BY, fdr, hom, hochberg, hommel, bonferroni, none}

\item{alternative}{two.sided, greater, less}

\item{conf_level}{confidence level for the returned confidence interval.}
}
\value{
Returns a dataframe.
x = x variable
y = y variable
group = sample group
n = number of samples in group
r = correlation coefficient
p = p-value
p.adj = adjusted p-value
lower_ci = lower confidence interval
upper_ci = upper confidence interval
}
\description{
Performs pairwise correlations of features with adjusted p-values.
Correlations and confidence intervals calculated for each sample group.
}
\examples{
# load example data
data(dss_example)

# subset features, features in columns
feat_mat <- dss_example$merged_abundance_id[1:6,2:26]
rownames(feat_mat) <- dss_example$merged_abundance_id[1:6,1]
feat_mat <- t(feat_mat)

# metadata in same order
met_df <- dss_example$metadata
met_df <- met_df[match(rownames(feat_mat), met_df$sampleID),]

compare_cor_ci(feat_mat, met_df$Phenotype)
}
