% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_platemap.R
\name{convert_platemap}
\alias{convert_platemap}
\title{convert_platemap}
\usage{
convert_platemap(map_file, sheet = NULL, map_range, drop_empty = TRUE, ...)
}
\arguments{
\item{map_file}{excel file containing plate map}

\item{sheet}{name or number of excel sheet. defaults to first sheet}

\item{map_range}{string. cell range containing plate map format
using excel notation (e.g. 'A1:H12')}

\item{drop_empty}{boolean. default TRUE. drop unlabeled wells}

\item{...}{Arguments passed to \code{readxl::read_excel}}
}
\value{
datafame of platemap in long form
}
\description{
function to convert excel plate map to long data frame with map locations.
uses \code{readxl} to read in excel file so additional arguments
are passed to \code{readxl::read_excel}
}
