% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity.R
\name{dissimilarity}
\alias{dissimilarity}
\title{dissimilarity}
\usage{
dissimilarity(
  abundance_matrix = NULL,
  metadata = NULL,
  individual_variable = NULL,
  method = "within"
)
}
\arguments{
\item{abundance_matrix}{dataframe; samples are columns and features are rows}

\item{metadata}{dataframe; samples are rows and rownames should match those in abundance_matrix}

\item{individual_variable}{character; name of the variable in metadata that specifies the individual}

\item{method}{character; choose 'within' or 'between' depending on whether there are paired samples available}
}
\value{
data frame
}
\description{
Calculate either within-individual or between-individual Bray-Curtis dissimilarity.
}
\examples{
abundance_matrix <- data.frame(matrix(rnorm(1000, mean=100, sd=1), ncol=8, nrow=50))
colnames(abundance_matrix) <- LETTERS[1:8]
abundance_matrix <- ocms_relab(abundance_matrix)
metadata <- data.frame(individual = c(rep("1", 2), rep("2", 2),
                                      rep("3", 2), rep("4", 2)))
metadata$other_variable <- rnorm(8)
rownames(metadata) <- LETTERS[1:8]
dissimilarity(abundance_matrix, metadata=metadata,
              individual_variable="individual", method="within")
}
