% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalette.R
\name{getPalette}
\alias{getPalette}
\title{getPalette}
\usage{
getPalette(n = NULL, palette = NULL, preview = FALSE)
}
\arguments{
\item{n}{numeric; default NULL. number of colours; cannot exceed 335. when NULL, gives all colours (in palette, if specified).}

\item{palette}{character; allows you to specify a RColorBrewer palette by name
Diverging
BrBG, PiYG, PRGn, PuOr, RdBu, RdGy, RdYlBu, RdYlGn, Spectral

\if{html}{\out{<div class="sourceCode">}}\preformatted{   Qualitative
   Accent, Dark2, Paired, Pastel1, Pastel2, Set1, Set2, Set3

   Sequential
   Blues, BuGn, BuPu, GnBu, Greens, Greys, Oranges, OrRd, PuBu,
   PuBuGn, PuRd, Purples, RdPu, Reds, YlGn, YlGnBu, YlOrBr, YlOrRd
}\if{html}{\out{</div>}}}

\item{preview}{logic; default FALSE; when set, gives preview of colour}
}
\value{
vector of HEX colours, if full set to TRUE, gives all 335 colours, else gives n numbers
}
\description{
generates colour palette from RColorBrewer
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{getPalette} was renamed to \code{get_palette} for more consistent
function naming nomenclature in \code{OCMSutility}. Usage is still exactly the same
}
\examples{
getPalette(n=5, palette = 'Set3')
# ->
get_palette(palette = c('Set1','Dark2'), preview = TRUE)
}
\keyword{internal}
