% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metfile_init.R
\name{metfile_init}
\alias{metfile_init}
\title{function to initiate a metadata file from dada2 db file}
\usage{
metfile_init(
  db_file,
  out_dir = NULL,
  ref_table = NULL,
  id_orient = "col",
  dummy = NULL
)
}
\arguments{
\item{db_file}{rsqlite database file}

\item{out_dir}{output directory. default NULL (no output written to file)}

\item{ref_table}{= name of table in database from which sampleID are generated.
default NULL. when NULL assumes table is merged_abundance_id
which is the count table produced from the ocms_16s dada2 pipeline}

\item{id_orient}{= indicates orientation of sample IDs. default 'col' indicates
samples are in columns. 'row' indicates samples are in rows and
sample IDs are the rownames}

\item{dummy}{= default NULL. when string supplied, adds a dummy column of NAs with the string as the column name.}
}
\value{
dataframe of metadata
}
\description{
metfile_init
}
