% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OCMSutility-package.R
\docType{package}
\name{OCMSutility-package}
\alias{OCMSutility}
\alias{OCMSutility-package}
\title{OCMSutility: Utility functions for OCMS}
\description{
Package holds functions that are commonly used in microbiome analysis. Part of the OCMS sandbox toolkit.
}
\author{
\strong{Maintainer}: Sandi Yen \email{sandi.yen@kennedy.ox.ac.uk}

Authors:
\itemize{
  \item Nicholas Ilott \email{nicholas.ilott@kennedy.ox.ac.uk}
}

}
\keyword{internal}
