% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCoA.R
\name{plotPCoA}
\alias{plotPCoA}
\title{plotPCoA}
\usage{
plotPCoA(relab, met, colour = NULL, shape = NULL, CI = 0.95)
}
\arguments{
\item{relab}{dataframe. relative abundance data with features in rows
and samples in columns. feature IDs in rowname.}

\item{met}{dataframe. metadata with samples in rows. sample IDs in rowname}

\item{colour}{string. defulat NULL. metadata variable to colour points by}

\item{shape}{string. default NULL. metadata variable to set shape of points by}

\item{CI}{numeric. Default 0.95. Confidence interval used to draw ellipse
around colour variable. set to NULL to omit drawing ellipse}
}
\description{
This is a simple PCoA function that colours all points by one
metadata variable. It can be helpful to visualise metadata variables
independently when assessing potential confounding metadtaa factors
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plotPCoA} was renamed to \code{plotPCoA} for more consistent
function naming nomenclature in \code{OCMSutility}. Usage is still exactly the same
}
\examples{
data(dss_example)
met_df <- dss_example$metadata

count_df <- dss_example$merged_abundance_id \%>\%
  column_to_rownames('featureID')
count_df <- count_df[,met_df$sampleID]
relab <- relab(count_df)

iter_var <- c('Genotype','Phenotype')
for(i in iter_var) {
  plotPCoA(relab, met_df, colour = i)
  # ->
  plot_pcoa(relab, met_df, colour = i)
}
}
\keyword{internal}
