% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_abundance.R
\name{prevalence_abundance}
\alias{prevalence_abundance}
\title{prevalence_abundance
return a dataframe that gives the prevalence and abundance (mean across samples)
of each taxon (from relative abundance matrix)}
\usage{
prevalence_abundance(abundance_matrix)
}
\arguments{
\item{abundance_matrix}{dataframe; samples are columns and features are rows (relative abundance)}
}
\value{
dataframe
}
\description{
prevalence_abundance
return a dataframe that gives the prevalence and abundance (mean across samples)
of each taxon (from relative abundance matrix)
}
\examples{
prevalence_abundance(abundance_matrix)
}
