% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{rarefaction}
\alias{rarefaction}
\title{rarefaction}
\usage{
rarefaction(df)
}
\arguments{
\item{df}{dataframe with samples in columns, features in rows}
}
\value{
list of ggplot rarefaction curve and associated dataframe
}
\description{
produce rarefaction plot.
adapted from MetaSequencingSnake enumeration_report.Rmd
}
\examples{
count_data <-  cbind(matrix(rnorm(100*5,mean=50000,sd=10000), 100, 5),
                     matrix(rnorm(100*5,mean=20000,sd=5000), 100, 5),
                     matrix(rnorm(100*5,mean=5000,sd=1000), 100, 5))
rarefaction(as.data.frame(count_data))
}
