% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotate_tax.R
\name{reannotate_tax}
\alias{reannotate_tax}
\title{reannotate_tax}
\usage{
reannotate_tax(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy table. Must have \code{c('Kingdom','Phylum','Class','Order','Family','Genus','Species')} as columns. unclassified entries are denoted as NA}
}
\value{
updated taxonomy table with "unclassified" prepended with higher level classifications
}
\description{
Reannotate taxonomy

Reannotates taxonomy table so that "unclassfied" assignments include higher
level classifications. This helps preserve the biological meaning of an unclassfied genus (as it could be classfied at the Family level).
}
\examples{
data(asv_example)
# adding Kingdom column; removing sequence column because don't need asv IDs in this example
old_tax <- tax_example
colnames(old_tax)[1] <- 'Kingdom'
old_tax$Kingdom <- 'Bacteria'
knitr::kable(head(old_tax))

new_tax <- reannotate_tax(old_tax)
knitr::kable(head(new_tax))
}
