% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relab.R
\name{relab}
\alias{relab}
\title{relab}
\usage{
relab(counts)
}
\arguments{
\item{counts}{dataframe; samples are columns and features are rows}
}
\value{
dataframe
}
\description{
convert counts table to relative abundance (\%)
}
\examples{
counts <- data.frame(matrix(rnorm(1000), ncol=20, nrow=50))
rel_abundance <- relab(counts)
}
