% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_remove.R
\name{remove_geom}
\alias{remove_geom}
\title{remove_geom}
\usage{
remove_geom(ggplot2_object, geom_type)
}
\arguments{
\item{ggplot2_object}{ggplot2 object}

\item{geom_type}{}
}
\description{
Remove geom layer from ggplot. Based on https://stackoverflow.com/questions/13407236/remove-a-layer-from-a-ggplot2-chart
}
\examples{
d <- data.frame(
x = runif(10),
y = runif(10),
label = sprintf("label\%s", 1:10))

# ggplot with geom_text_repel from ggrepel
p <- ggplot(d, aes(x, y, label = label)) +
    geom_point() +
    geom_text()

# Remove the labels added by ggrepel.
p <- remove_geom(p, "GeomText")
}
