% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_barchart.R
\name{stacked_barchart}
\alias{stacked_barchart}
\title{stacked_barchart
stacked barchart of the top_n features in a relative abundance table}
\usage{
stacked_barchart(abundance_matrix, top_n = 10)
}
\arguments{
\item{abundance_matrix}{dataframe; samples are columns and features are rows (relative abundance)}

\item{top_n}{integer; number of taxa to plot in the stacked barchart. Default is 10.}
}
\value{
grob
}
\description{
stacked_barchart
stacked barchart of the top_n features in a relative abundance table
}
\examples{
stacked_barchart()
}
