% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_mat2df.R
\name{sym_mat2df}
\alias{sym_mat2df}
\title{sym_mat2df}
\usage{
sym_mat2df(mat)
}
\arguments{
\item{mat}{symmetrical matrix}
}
\value{
long dataframe with columns x, y, and value
}
\description{
Convert symmetrical matrix to long dataframe.
Helpful for correlation or distance matrices
}
\examples{
# load example data
data(dss_example)

# subset features, features in columns
feat_mat <- dss_example$merged_abundance_id[1:6,2:26]
rownames(feat_mat) <- dss_example$merged_abundance_id[1:6,1]
feat_mat <- t(feat_mat)

# correlation matrix
corr_result <- cor(feat_mat)

sym_mat2df(corr_result)

}
