% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zymobiomics.R
\docType{data}
\name{zymobiomics}
\alias{zymobiomics}
\title{expected abundance and annotation for zymobioimics samples}
\format{
list of 4 dataframes
}
\usage{
data(zymobiomics)
}
\description{
Accompanies asv_example, which is a subset of data from a V3-V4 16S rRNA human dataset
}
\keyword{datasets}
