% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateCount.R
\name{aggregateCount}
\alias{aggregateCount}
\title{aggregateCount}
\usage{
aggregateCount(count_df, tax_df = NULL, aggregate_by = NULL)
}
\description{
Aggregate read counts based on taxonomic level
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{aggregateCount} was renamed to \code{aggregate_count} for more consistent
function naming nomenclature in \code{OCMSutility}. Usage is still exactly the same
}
\examples{
data(dss_example)
# featureID should be row names
feature_count <- dss_example$merged_abundance_id \%>\%
  tibble::column_to_rownames('featureID')

# cleanup sample names
colnames(feature_count) <- paste0('id', colnames(feature_count))
feature_tax <- dss_example$merged_taxonomy

# set row order of count and tax tables to be the same
feature_count <- feature_count[feature_tax$featureID,]

aggregated_list <- aggregateCount(feature_count, feature_tax,
                                  aggregate_by = "Family")
# ->
aggregated_list <- aggregate_count(feature_count, feature_tax,
                                  aggregate_by = "Family")

summary(aggregated_list)
}
\keyword{internal}
