% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R, R/plot_pca.R
\name{getVE}
\alias{getVE}
\title{getVE}
\usage{
getVE(pc, component = "PC1")

getVE(pc, component = "PC1")
}
\arguments{
\item{pc}{prcomp object}

\item{component}{principal component}
}
\value{
matrix of variance explained for a principal component

matrix of variance explained for a principal component
}
\description{
Calculate variance explained in pca. used in plotPrincipalComponents

Calculate variance explained in pca. used in plotPrincipalComponents
}
