% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotateTax.R
\name{reannotateTax}
\alias{reannotateTax}
\title{reannotateTax}
\usage{
reannotateTax(taxonomy)
}
\arguments{
\item{taxonomy}{taxonomy table. Must have \code{c('Kingdom','Phylum','Class','Order','Family','Genus','Species')} as columns. unclassified entries are denoted as NA}
}
\value{
updated taxonomy table with "unclassified" prepended with higher level classifications
}
\description{
Reannotate taxonomy
`r lifecycle::badge("deprecated")`

`reannotateTax` was renamed to `reannotate_tax` for more consistent
function naming nomenclature in `OCMSutility`. Usage is still exactly the same
}
\examples{
data(asv_example)
# adding Kingdom column; removing sequence column because don't need asv IDs in this example
old_tax <- tax_example
colnames(old_tax)[1] <- 'Kingdom'
old_tax$Kingdom <- 'Bacteria'
knitr::kable(head(old_tax))

new_tax <- reannotateTax(old_tax)
# ->
new_tax <- reannotate_tax(old_tax)
knitr::kable(head(new_tax))
}
\keyword{"unclassfied"}
\keyword{(as}
\keyword{Family}
\keyword{Reannotates}
\keyword{This}
\keyword{an}
\keyword{assignments}
\keyword{at}
\keyword{be}
\keyword{biological}
\keyword{classfied}
\keyword{classifications.}
\keyword{could}
\keyword{genus}
\keyword{helps}
\keyword{higher}
\keyword{include}
\keyword{internal}
\keyword{it}
\keyword{level}
\keyword{level).}
\keyword{meaning}
\keyword{of}
\keyword{preserve}
\keyword{so}
\keyword{table}
\keyword{taxonomy}
\keyword{that}
\keyword{the}
\keyword{unclassfied}
